// frontend/src/components/layout/Header/Header.types.ts
/**
 * ===============================================================
 * Author: ultramegared
 * Project: Magic Touch Designs
 * File: Header.types.ts
 * Module: Frontend
 * Language: TypeScript
 * Description:
 * Header navigation types.
 * ===============================================================
 */

export interface NavigationItem {

    label: string;

    path: string;

}